refactor(frontend): extract useTableActions composable (P2-2 F15)
Table actions and detail loaders (~288 lines) move from DataStudioNew.vue
into composables/useTableActions.js — saveAsTask, handleCreateTable,
handleDeleteTable, syncMissingTableMetadata, handleCreateSuccess, loadDdl,
loadAccessStats, copyDdl, openTask, goCreateRelatedTask, handleTaskSuccess,
goLineage. The metaTab watcher stays in the component and now calls the
composable-returned loaders; provide('dataStudioCtx') keys unchanged.
Imports that moved with the code (ElMessageBox, lineageApi, copyText,
showDemoReadonlyMessage) are dropped from the component.

New __tests__/useTableActions.spec.js covers the loadDdl id/by-name paths
and failure handling, loadAccessStats dedupe/cache/error states, the
delete confirmation and cancel flows, metadata sync, goLineage guard and
routing, saveAsTask validation, copyDdl, and handleTaskSuccess (12 tests).

DataStudioNew.vue: 2380 -> 2122 lines.

Verified: npm run lint (0 errors), npm run test (29 files / 146 passed,
incl. DataStudioNew mount smoke), npm run build; demo-mode Playwright
smoke (no real backend): open table tab -> DDL tab loads real demo DDL,
access tab renders, new query tab executes SELECT 1 end-to-end.
